z_keyexpr_unchecked, z_query_value covered
