+++ include/zenohcpp_objects.h
@@ -44,28 +44,6 @@ class KeyExpr : public Owned<::z_owned_keyexpr_t> {
 KeyExpr KeyExprView::concat(const std::string_view& s) const { return ::z_keyexpr_concat(*this, s.data(), s.length()); }
 KeyExpr KeyExprView::join(const KeyExprView& v) const { return ::z_keyexpr_join(*this, v); }
 
-bool keyexpr_canonize(std::string& s, ErrNo& error) {
-    uintptr_t len = s.length();
-    error = ::z_keyexpr_canonize(&s[0], &len);
-    s.resize(len);
-    return error == 0;
-}
-
-bool keyexpr_canonize(std::string& s) {
-    ErrNo error;
-    return keyexpr_canonize(s, error);
-}
-
-bool keyexpr_is_canon(const std::string_view& s, ErrNo& error) {
-    error = ::z_keyexpr_is_canon(s.begin(), s.length());
-    return error == 0;
-}
-
-bool keyexpr_is_canon(const std::string_view& s) {
-    ErrNo error;
-    return keyexpr_is_canon(s, error);
-}
-
 class ScoutingConfig;
 
 class Config : public Owned<::z_owned_config_t> {
